Add Robots tests to validate nginx behaviour.
- nginx adds X-Robots-Tag depending on the domain to restrict crawlers.
- nginx delivers a robots disallow based on host.
