+++ .docker/Dockerfile.nginx-drupal
@@ -6,6 +6,7 @@ FROM amazeeio/nginx-drupal:${LAGOON_IMAGE_VERSION}
 
 # nginx config.
 COPY .docker/images/nginx/helpers/ /etc/nginx/helpers/
+
 COPY .docker/images/nginx/no-robots.sh /lagoon/entrypoints/20-no-robots.sh
 RUN fix-permissions /etc/nginx
