+++ .docker/images/nginx/helpers/200-no_robots_govcms_host.conf
@@ -6,7 +6,6 @@ if ($uri = "/robots.txt"){
 # matches *.govcms.gov.au (also multiple subdomains) except www.govcms.gov.au
 if ($host ~* ^(?!www\.govcms\.gov\.au$)(?:[\w\-]+\.)+govcms\.gov\.au$){
   set $robots_disallow_test "$robots_disallow_test+isGovcmsHost";
-  set $govcms_host "true";
 }
 
 if ($robots_disallow_test = "isRobotstxtURI+isGovcmsHost"){
