+++ .docker/images/nginx/helpers/206-x_robots_tag_header.conf
@@ -1,0 +1,14 @@
+###
+### Add the X-Robots-Tag to the server response
+### @see https://developers.google.com/search/reference/robots_meta_tag
+###
+
+# All is default and essentially has no effect.
+set $robots_header "all";
+
+if ($host ~* ^(?!www\.govcms\.gov\.au$)(?:[\w\-]+\.)+govcms\.gov\.au$){
+  # None = noindex, nofollow.
+  set $robots_header "none";
+}
+
+add_header X-Robots-Tag $robots_header;
